+++ addons/purchase_stock/__manifest__.py
@@ -19,6 +19,7 @@
         'views/res_config_settings_views.xml',
         'views/res_partner_views.xml',
         'views/stock_lot_views.xml',
+        'views/product_views.xml',
         'report/purchase_report_views.xml',
         'report/purchase_report_templates.xml',
         'report/report_stock_rule.xml',
