+++ addons/purchase_stock/models/account_invoice.py
@@ -42,8 +42,14 @@ class AccountMove(models.Model):
                     continue
 
                 # Retrieve accounts needed to generate the price difference.
-                debit_expense_account = line._get_price_diff_account()
-                if not debit_expense_account:
+                debit_pdiff_account = False
+                if line.product_id.cost_method == 'standard':
+                    debit_pdiff_account = line.product_id.property_account_creditor_price_difference \
+                        or line.product_id.categ_id.property_account_creditor_price_difference_categ
+                    debit_pdiff_account = move.fiscal_position_id.map_account(debit_pdiff_account)
+                else:
+                    debit_pdiff_account = line.product_id.product_tmpl_id.get_product_accounts(fiscal_pos=move.fiscal_position_id)['expense']
+                if not debit_pdiff_account:
                     continue
                 # Retrieve stock valuation moves.
                 valuation_stock_moves = self.env['stock.move'].search([
@@ -73,7 +79,6 @@ class AccountMove(models.Model):
                         price_unit, move.currency_id,
                         move.company_id, valuation_date, round=False
                     )
-
 
                 price_unit = line._get_gross_unit_price()
 
@@ -110,7 +115,7 @@ class AccountMove(models.Model):
                             line.company_currency_id,
                             line.company_id, fields.Date.today(),
                         ),
-                        'account_id': debit_expense_account.id,
+                        'account_id': debit_pdiff_account.id,
                         'analytic_distribution': line.analytic_distribution,
                         'display_type': 'cogs',
                     }
