+++ addons/purchase_stock/models/product.py
@@ -5,9 +5,23 @@ from odoo import api, fields, models
 from odoo.osv import expression
 
 
+class ProductCategory(models.Model):
+    _inherit = "product.category"
+
+    property_account_creditor_price_difference_categ = fields.Many2one(
+        'account.account', string="Price Difference Account",
+        company_dependent=True,
+        help="This account will be used to value price difference between purchase price and accounting cost.")
+
+
 class ProductTemplate(models.Model):
     _name = 'product.template'
     _inherit = 'product.template'
+
+    property_account_creditor_price_difference = fields.Many2one(
+        'account.account', string="Price Difference Account", company_dependent=True,
+        help="This account is used in automated inventory valuation to "\
+             "record the price difference between a purchase order and its related vendor bill when validating this vendor bill.")
 
     @api.model
     def _get_buy_route(self):
