+++ addons/purchase_stock/views/product_views.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <record id="view_category_property_form" model="ir.ui.view">
-        <field name="name">product.category.property.form.inherit.stock</field>
+        <field name="name">product.category.property.form.inherit.purchase.stock</field>
         <field name="model">product.category</field>
         <field name="inherit_id" ref="account.view_category_property_form"/>
         <field name="arch" type="xml">
@@ -14,7 +14,7 @@
     </record>
 
     <record id="product_template_form_view" model="ir.ui.view">
-        <field name="name">product.normal.form.inherit.stock</field>
+        <field name="name">product.normal.form.inherit.purchase.stock</field>
         <field name="model">product.template</field>
         <field name="inherit_id" ref="account.product_template_form_view"/>
         <field name="arch" type="xml">
